Create Directory and Add OSX Flag
When using this SCons file the build stops if there is no 'demo/bin' directory. A change was made so the directory is created before we update the environment. -- "scons: *** Path for option target_path does not exist: demo/bin/"

Added c++17 flags for OSX platform so build would not fail when using c++11 features. -- "warning: scoped enumerations are a C++11 extension [-Wc++11-extensions]"